Entity 수정/ 거래 조건 추가

+++ quant-core/src/main/java/com/quant/core/entity/Portfolio.java
@@ -27,7 +27,7 @@ public class Portfolio {
     UserInfo userInfo;
 
     //전체 투자금
-    Integer totalValue;
+    Long totalValue;
 
     //최소 모멘텀
     Integer momentumScore;
